chore(orchestrator-backend): trim unused production exports and dead symbols - orchestrator-backend
