+++ workspaces/orchestrator/plugins/orchestrator-backend/src/service/Helper.ts
@@ -42,30 +42,6 @@ export async function retryAsyncFunction<T>(args: {
   throw new Error('Exceeded maximum number of retries for async function');
 }
 
-export async function getWorkingDirectory(
-  config: Config,
-  logger: LoggerService,
-): Promise<string> {
-  if (!config.has('backend.workingDirectory')) {
-    return os.tmpdir();
-  }
-
-  const workingDirectory = config.getString('backend.workingDirectory');
-  try {
-    // Check if working directory exists and is writable
-    await fs.access(workingDirectory, fs.constants.F_OK | fs.constants.W_OK);
-    logger.info(`using working directory: ${workingDirectory}`);
-  } catch (err: any) {
-    logger.error(
-      `working directory ${workingDirectory} ${
-        err.code === 'ENOENT' ? 'does not exist' : 'is not writable'
-      }`,
-    );
-    throw err;
-  }
-  return workingDirectory;
-}
-
 export async function executeWithRetry(
   action: () => Promise<Response>,
   maxErrors = 15,
