+++ workspaces/orchestrator/plugins/orchestrator-backend/src/service/api/test-utils.ts
@@ -17,6 +17,7 @@
 import { DateTime } from 'luxon';
 
 import {
+  PaginationInfoDTO,
   ProcessInstance,
   ProcessInstanceState,
   ProcessInstanceStateValues,
@@ -28,7 +29,42 @@ import {
   WorkflowOverviewListResult,
 } from '@red-hat-developer-hub/backstage-plugin-orchestrator-common';
 
+import { Pagination } from '../../types/pagination';
+
 const BASE_DATE = '2023-02-19T11:45:21.123Z';
+
+export function buildPaginationTmp(
+  paginationInfo?: PaginationInfoDTO,
+): Pagination {
+  const pagination: Pagination = {
+    limit: undefined,
+    offset: undefined,
+    order: undefined,
+    sortField: undefined,
+  };
+
+  if (!paginationInfo) {
+    return pagination;
+  }
+  const { offset, pageSize, orderBy, orderDirection } = paginationInfo;
+
+  if (!isNaN(Number(offset))) {
+    pagination.offset = Number(offset);
+  }
+
+  if (!isNaN(Number(pageSize))) {
+    pagination.limit = Number(pageSize);
+  }
+
+  if (orderBy) {
+    pagination.sortField = String(orderBy);
+  }
+
+  if (orderDirection) {
+    pagination.order = String(orderDirection).toUpperCase();
+  }
+  return pagination;
+}
 
 interface WorkflowOverviewParams {
   suffix?: string;
