+++ workspaces/orchestrator/plugins/orchestrator-backend/src/service/constants.ts
@@ -16,5 +16,4 @@
 
 export const WORKFLOW_DATA_KEY = 'workflowdata';
 
-export const INTERNAL_SERVER_ERROR_MESSAGE = 'internal server error';
 export const FETCH_PROCESS_INSTANCES_SORT_FIELD = 'start';
